feat(apple login): Enhance Auth with Apple Sign-In & update IAM policies
Integrated Sign-In with Apple as an external provider in the auth configuration. Added support for callback and logout URLs. Disabled unauthenticated identities in the Cognito identity pool for enhanced security. Temporarily commented out unused code in the Authenticator component and utility function to focus on the current authentication setup.

Addresses security and usability improvements effectively.